refactor: 회원가입 부분 상태 관리 폴더 변경
