+++ src/pages/register/store/type/index.ts
